Fix reference count for return value of LRU.setdefault().
In commit a90f68fe, the setdefault() method was implemented. However,
the reference to the return value is stolen by the function and can be
lost when the referent is released. This is not the intended behaviour
and can crash Python because of garbage collection.

This is fixed by donating reference to the caller.