Implement the popitem() method.
The method `popitem()` follows the semantics of those of Python's `dict`
or `collections.OrderedDict`. An optional keyword argument,
`least_recent`, default to `True`, controls the order of removal. If
`True`, the least-recently used key-value pair is removed and returned.
If `False`, it is the most-recently used pair that is removed and
returned.