Add lock around the execution of the callback.
If a callback is set, cache eviction is guarded by a lock. In this case,
because callback will be executed and it may execute arbitrary code --
which may include the dropping of the Python GIL, -- the cache may
become corrupted due to racing. The lock added in this commit ensures
that access to the state of the cache is synced even if the callback may
drop the GIL.