Add mutex protection of internal critical section.
A mutex (Python lock) is added for the protection of the internal state.
Access to the internal data by the call to a Python-facing method is
done in a single pass through the critical section.

The callback, in contrast, will only be executed outside the critical
section. Inside the critical section, if a callback is set, cache
eviction will transfer the evicted objects into a staging area without
directly triggering the callback. Then, after exiting the critical
section, the staging area is purged by the callback.

In addition, some bugfixes are done.